refactor(tests): extract _ensure_discord_mock to tests/tools/conftest.py
Removes ~140 lines of near-identical _ensure_discord_mock duplication
between tests/tools/test_discord_button_message.py and
tests/tools/test_discord_reaction_tool.py. Both files now rely on a
shared helper at tests/tools/conftest.py that pytest auto-loads at
collection time before any test module in the same directory imports.

The reason the duplication existed in the first place is that
tests/gateway/conftest.py builds its own discord mock with
`discord_mod = MagicMock()` (fresh) and overwrites sys.modules,
clobbering attributes set by an earlier conftest. To make the two
conftests cooperate regardless of xdist worker collection order,
this commit also adds three lines to tests/gateway/conftest.py
(`Forbidden`, `NotFound`, `HTTPException` as real Exception
subclasses), so production isinstance checks in
tools/discord_button_tool.py and tools/discord_reaction_tool.py
work whichever conftest's mock ends up in sys.modules at runtime.

Verification:

* tests/tools/test_discord_button_message.py: 11 passed (parallel + serial)
* tests/tools/test_discord_reaction_tool.py: 13 passed (parallel + serial)
* Combined: 24 passed (parallel + serial)
* Full Track-1 suite (button + reaction + skill_resolver +
  discord_interactions + discord_inbound_reactions +
  skill_utils_triggers): 86 passed (parallel + serial)
* Broader gateway-discord subset (5 test files): 92 passed (no
  regressions in adjacent gateway tests)
* Net diff: -156 + 13 + 100 = -43 lines

Single source of truth for the discord mock shape, eliminating the
drift risk noted in cycle-3 architect review of commit 9.
