+++ tests/tools/test_discord_reaction_tool.py
@@ -26,83 +26,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import pytest
 
 
-# ---------------------------------------------------------------------------
-# discord mock (mirrors _ensure_discord_mock from test_discord_button_message.py)
-# ---------------------------------------------------------------------------
-
-def _ensure_discord_mock() -> None:
-    """Ensure sys.modules['discord'] has real Exception classes for NotFound/Forbidden/HTTPException.
-
-    Idempotent and resilient to other test files setting up bare-MagicMock discord modules
-    earlier in collection (xdist test order is non-deterministic). Real discord.py is left
-    alone if it's importable.
-    """
-    real_discord = sys.modules.get("discord")
-    if real_discord is not None and hasattr(real_discord, "__file__"):
-        return  # real discord.py installed — leave it alone
-
-    discord_mod = real_discord if real_discord is not None else MagicMock()
-
-    def _is_real_exception(obj: Any) -> bool:
-        return isinstance(obj, type) and issubclass(obj, BaseException)
-
-    if not _is_real_exception(getattr(discord_mod, "Forbidden", None)):
-        discord_mod.Forbidden = type("Forbidden", (Exception,), {})
-    if not _is_real_exception(getattr(discord_mod, "HTTPException", None)):
-        discord_mod.HTTPException = type("HTTPException", (Exception,), {})
-    if not _is_real_exception(getattr(discord_mod, "NotFound", None)):
-        discord_mod.NotFound = type("NotFound", (Exception,), {})
-
-    # NOTE: bare `MagicMock()` returns truthy auto-attrs for any name, so
-    # `hasattr(mm, "ui")` is always True. Use stricter type checks to detect
-    # whether a previous test (or the real discord.py) actually set the field.
-    if not isinstance(getattr(discord_mod, "DMChannel", None), type):
-        discord_mod.DMChannel = type("DMChannel", (), {})
-    if not isinstance(getattr(discord_mod, "Thread", None), type):
-        discord_mod.Thread = type("Thread", (), {})
-    if not isinstance(getattr(discord_mod, "ForumChannel", None), type):
-        discord_mod.ForumChannel = type("ForumChannel", (), {})
-    if getattr(discord_mod, "Interaction", None) is not object:
-        discord_mod.Interaction = object
-    if not isinstance(getattr(discord_mod, "ButtonStyle", None), SimpleNamespace):
-        discord_mod.ButtonStyle = SimpleNamespace(
-            primary="primary",
-            secondary="secondary",
-            success="success",
-            danger="danger",
-        )
-
-    ui_mod = getattr(discord_mod, "ui", None)
-    if not (isinstance(ui_mod, SimpleNamespace) and isinstance(getattr(ui_mod, "View", None), type)):
-        class _FakeView:
-            def __init__(self, *, timeout: float = 180.0) -> None:
-                self.timeout = timeout
-                self.children: list = []
-
-            def add_item(self, item: Any) -> None:
-                self.children.append(item)
-
-        class _FakeButton:
-            def __init__(self, *, label: str, custom_id: str, style: Any = "primary") -> None:
-                self.label = label
-                self.custom_id = custom_id
-                self.style = style
-                self.callback: Any = None
-
-        discord_mod.ui = SimpleNamespace(View=_FakeView, Button=_FakeButton)
-
-    sys.modules["discord"] = discord_mod
-    if "discord.ext" not in sys.modules:
-        ext_mod = MagicMock()
-        commands_mod = MagicMock()
-        commands_mod.Bot = MagicMock
-        ext_mod.commands = commands_mod
-        sys.modules["discord.ext"] = ext_mod
-        sys.modules["discord.ext.commands"] = commands_mod
-
-
-_ensure_discord_mock()
-
+# Discord mock is set up in tests/tools/conftest.py before this module imports.
 # Import after mock is in place
 from tools.discord_reaction_tool import (  # noqa: E402
     _handler_add_async,
